% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flexFNnew.R
\name{estFNsegs}
\alias{estFNsegs}
\title{Segment estimates}
\usage{
estFNsegs(eList, dateInfo)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{segStart_i}{integer vector of start years (water) for each FN conc/flux segment}
}
\description{
Segment estimates
}
\examples{
library(EGRET)
eList <- Choptank_eList
eList <- setUpEstimation(eList)
sampleSegStart <- c(1980,1990,2000)
flowSegStart <- c(1980,1985,1992)
flowSegEnd <- c(1994,2004,2011)
eList <- estFNsegs(eList,dateInfo[1,])
}

