% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{explore_endpoints}
\alias{explore_endpoints}
\title{Explore EGRET}
\usage{
explore_endpoints(browse = TRUE)
}
\arguments{
\item{browse}{use browser for map rendering}
}
\description{
Open an interactive app
}

