% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flexFNnew.R
\name{flexFN}
\alias{flexFN}
\title{Flexible Flow Normalization}
\usage{
flexFN(eList, sampleSegStart, flowSegStart, flowSegEnd)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{sampleSegStart}{integer vector of start years (water) for each FN conc/flux segment}

\item{flowSegStart}{integer vector of start years (water) for flow normalization}

\item{flowSegEnd}{integer vector of end years (water) for flow normalization}
}
\description{
Flexible Flow Normalization
}
\examples{
library(EGRET)
eList <- Choptank_eList
eList <- setUpEstimation(eList)
sampleSegStart <- c(1980,1990,2000)
flowSegStart <- c(1980,1985,1992)
flowSegEnd <- c(1994,2004,2011)
eList <- flexFN(eList,sampleSegStart, flowSegStart, flowSegEnd)
plotFluxHist(eList)
}

